apps/backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Database (optional - install separately if needed)
# sqlalchemy>=2.0.0
# alembic>=1.10.0
# psycopg2-binary>=2.9.0

# Redis and Task Queue (optional)
# redis>=5.0.0
# celery>=5.3.0

# Image processing (optional - install separately if needed)
# pillow>=10.0.0
# opencv-python>=4.8.0
# numpy>=1.24.0

# HTTP clients
requests>=2.31.0
httpx>=0.25.0

# AI Image Generation
google-genai>=0.8.0
pillow>=10.0.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
